refactor: extract spec comparison into specDiff.ts
isEqual becomes compareSpecifications(spec, other); isEqualValue is
exported alongside. The class keeps instance delegates.
