+++ backend/src/specification/index.ts
@@ -5,6 +5,7 @@ export * from './ifilespecification.js'
 export * from './m2mspecification.js'
 export * from './modbusValues.js'
 export * from './specMessages.js'
+export * from './specDiff.js'
 export * from './m2mgithub.js'
 export * from './configspec.js'
 export * from './migrator.js'
